Friday Changes
What I added today was I made it so that the player and the dealer's cards are drawn at the start of the game as in blackjack 2 random cards are drawn at the start of every game. I also made it to where the dealer will only stand at 17 or higher and will not stand before that unless the player is already standing. I also added in the logic that fixed it to where if the player were to stand at a number lower than the dealer it would result in the dealer still drawing more cards and losing despite winning already. I also added better text elements to make the text stand out in the different outcomes in the game for example losing the game will result in red text and winning will result in the green text I feel like this adds to the feeling of accomplishment.