Added in split feature
If you have two of the same face value cards for example a 7 and a 7 you can have the option to split the deck into another deck giving you 2 decks to play with and up to 3 decks can be split for a total of 4 decks at once. If you want to split the deck just click x on the keyboard when prompted. This feature is not yet finished as it doesn't fully alternate between the split pile and the original pile and if you stand or go over 21 in one pile of cards the game ends up not letting you play off the second pile that isn't at 21 and hasn't stood. Another issue is when you have the option to split a pile and you don't want to split the text for splitting still to appear on the screen which it shouldn't.