Added reset feature
Added in reset feature that will reset the game when clicking R on the keyboard. It will reset all card elements on screen and reshuffle the deck so when you restart the game everytime the deck is random. I also fixed an issue where some cards were able to split despite no having the same face value as each other. So now you can only split your deck if the cards are matching in face value.