Bug fixes
I noticed if both the player and dealer are at 21 it leads to a win for the player regardless of the dealer having the exact score. I fixed this issue and now if both the dealer and the user are at 21 it results in a tie. I also noticed if the player hit 21 the game would automatically end giving the player the win despite the dealer being at a score lower than 17 and not standing. I made it to where now when you reach a score of 21 standing or not the dealer will draw cards if their score is lower than 17 to try and match the user's score of 21.