Added betting system
Added in betting feature so that you can now bet in increments of 50 each increment of 50 that is bet will have a coin object drawn for it so for example if you bet 300 chips you will get 6 coins on the blackjack table each being 50 a piece. The initial state of the game was also updated to include a betting state that begins before the playing state and asks the player how much they want to bet. In order to bet you have to be at least 50 chips to play the game won't let you progress if you do not select the number of chips you can use the left arrow and right arrow to change the amount of chips you want to bet. Once you have the amount of chips you want to bet you can now press enter which will lock in your bet and set the game to its playing state where you can carry out a game of blackjack. There are also win-and-lose conditions to the betting for example if you bet 200 chips and lose you will lose those chips and your total amount of chips left will go down I set the starting amount of chips at 1000 chips so if you bet 200 and lose you will only have 800 chips left to bet but if you win and bet 200 chips you will get your 200 chips back plus another 200 giving you a score of 1200 chips.