Flipped Dealers First Card
I flipped the dealer's first card to be more accurate in the game of blackjack as in the game the dealer's first card is always flipped until near the end and the player's turn is over. I added in the sprite which shows the back of the card and after the player's turn is up and it's the dealer's turn to draw the card will flip itself over showing its value then will result in a win loss or draw of the player or dealer. The dealer also knows the value of the card so it stops them from overdrawing cards or having an incorrect balance.