Added in goal for the game
What I did was add in a goal for the game since the game is blackjack I thought It would be a good idea to make the goal of the level that the player has to hit a certain amount of chips to win. Right now the game will ask you upon start how much you would like to bet out of your starting amount which is 1k chips. After the player selects the amount of chips they would like to bet they are met with a goal that requires them to get a certain amount of chips in our case right now it's set to 2k chips to beat the level just to make it easier for testing. If the player hits 2k chips the game will transition into a sprite image which would be our win screen but in this case, is replaced with an arrow sprite. After the player sees the win screen they are then prompted to click enter on the keyboard to continue. Upon continuing the game will take you to the main menu/another room right now I have it set to go to the game room but it can be changed. I have implemented the same function for when you lose the game and it behaves the same way it just has an issue where the game is ending then after it shows the win screen for a few seconds it goes back to showing the game. Also, I forgot to make it to where the win screen triggers if you bust on 0 chips so as of right now if you bust and have no chips left the screen won't show it will only show if you have the lose condition. I also made a progress bar that will fill up as you get more and more chips as well as an on-screen prompt that lets you know the current goal of the level. I also moved some text elements around to make the UI look nicer.