Added time limit and win condition for busting
What I added in this version was a timer to the level that will give the level some pace and steaks making it feel more like a game. The time was added right now the timer is 60 seconds but can be adjusted later. The timer starts after the betting stage ends and the player transitions into the playing state. The timer also has a countdown bar that counts down as the time is going down so you get 2 visual representations of the time one in numbers and the other in the form of a bar that's steadily decreasing. Once the time is up it will tell the player that the time has expired and to press Enter to retry the level upon pressing enter the game room will reload letting the player play again. The time limit has also been made to reset every time the room is re-loaded so if you click R on the keyboard to reset the game the timer will also. Keep in mind the time does still count down as you are trying to get to the chip goal of 2k. I also added in a win condition for busting similar to how it is for winning and losing when the player is out of chips now appropriately when the player is out of chips and either bust wins or goes over 21 a corresponding text prompt will pop up telling you to click enter to continue once you click enter it will transition the player into the game room (which will be changed later) where an image of some sort will show up.