Added time increase and decrease mechanic
If the player wins a hand to the dealer the time gained to the timer is 10 seconds if the player loses or busts the time lost on the timer is 10 seconds. If the player and the dealer draw nothing happens and the timer continues as normal. There is also a visual feedback that is shown above the screen that will say +10s or -10s to let the player know they have lost or gained time. The timer will also adjust to the changes.