Added max bet and fixed issues
There is now a max bet of 250 chips for the level, making it so that the game cannot be completed in one turn by betting all chips and winning. Also added text on the screen that lets the player know the max bet is 250 chips. You can no longer reset the game after the timer is up forcing the player to the game over screen correctly. Fixed timer issue where the timer would sometimes run off-screen.