Updates
The coins are now stacked in a horizontal manner on the right side of the screen instead of the middle as It would conflict with text elements. I also updated most of the text elements to have some type of background to make it stand out in comparison to the green blackjack table. Sound effects were also added, such as a coin sound effect that plays when chips are drawn to the table. There is also another sound for the coins when you lose a hand that sounds different. There are now also 2 bell sound effects that were added to the game one for when the player wins and gains 10 seconds and one for when the player loses or busts and loses 10 seconds. There is now a sound for winning that plays once the player hits 2k chips and beats the level. Moved more elements around to make the game room look better.