Sound and Text Elements Update
What I have added in this update is a time-expired sound effect that is played when the time is up. The timer sound stops after the sound and resumes when the round starts again and resters when clicking R. I also added a sound effect when fully losing or busting out of chips similar to how it is when you get all the chips in the level. I added in a sound effect when the player and the dealer tie during a hand. When you now win a hand a coin sound effect is played similar to the sound effect when you lose a hand. I added more backgrounds behind more text to make the text stand out from the green background of our game room for text elements like winning losing and busting fully in the level in the level. Also changed some text to be better worded. There is now a sound effect when increasing and decreasing the bet.